Refactor IndoorNavigation components to use BuildingDataService and remove unused FloorPlanService and FloorNavigation
