+++ ConcordiaMaps/components/IndoorNavigation/FloorSelector.js
@@ -3,7 +3,7 @@ import { View, Text, ScrollView, TouchableOpacity, StyleSheet, Dimensions } from
 import { useNavigation, useRoute } from '@react-navigation/native';
 import Header from '../Header';
 import NavBar from '../NavBar';
-import FloorRegistry from './FloorRegistry';
+import FloorRegistry from '../../services/BuildingDataService';
 
 
 const FloorSelector = () => {
